Buildfixes

+++ Core/Util/TimeStretcher.cpp
@@ -103,7 +103,7 @@ namespace AudioCore {
 	}
 
 	int TimeStretcher::GetSamplesQueued() {
-		return impl->samples_queued;
+		return (int)impl->samples_queued;
 	}
 
 	void TimeStretcher::Flush() {
